chore(benches): extract the shared runtime helpers (4/5)
Hashing every freestanding helper body across the 15 targets found exactly one
duplicate: `runtime()`, in two byte-identical clusters. `make_document`,
`bulk_load` and `insert` share names across targets but have different bodies,
so they stay where they are.

`common/mod.rs` was db's vector fixtures rather than a shared root; it
becomes `common/vector.rs`, and the new `common/mod.rs` holds only what more
than one target uses:

  shared_runtime()  &'static OnceLock  blockstore, groupby, merge, transaction
  owned_runtime()   fresh per call     sift, vector_search, vector_mutations

The four remaining variants stay with their targets, and the module doc says
why. Their builder configuration is part of what they measure: http's runtime
must be multi-threaded or the slow path deadlocks on `Handle::block_on` inside
`spawn_blocking`; crdt runs a custom single-poll executor alongside
`Runtime::block_on` precisely so the gap between them can be measured. Merging
those would be a rewrite, not a move.

  cargo bench --no-run --workspace                     exit 0, 15 targets
  cargo clippy -p benches --all-targets -D warnings    exit 0
  criterion group names unchanged, no string in two binaries
